discord-bot/pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot for team management and ClickUp integration"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "supabase>=2.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.uv]
dev-dependencies = []
